Rework som unit tests
Move start date time tests for different kinds of groups to their own test files
